refactor: extract Modal
