+++ packages/apps/board-vitesse/src/components.d.ts
@@ -16,6 +16,7 @@ declare module 'vue' {
     Footer: typeof import('./components/Footer.vue')['default']
     GirlIcon: typeof import('./components/icon/GirlIcon.vue')['default']
     GoBack: typeof import('./components/GoBack.vue')['default']
+    Modal: typeof import('./components/board/Modal.vue')['default']
     NavBar: typeof import('./components/NavBar.vue')['default']
     Progress: typeof import('./components/board/Progress.vue')['default']
     RightArrowIcon: typeof import('./components/icon/RightArrowIcon.vue')['default']
